Created a function that shows the modal in modern way when edit button is clicked

+++ pages/details/[id].js
@@ -1,12 +1,16 @@
-import Header from "@/components/Header";
+
 import { useRouter } from "next/router";
 import TransactionDetails from "@/components/TransactionDetails";
-import * as Styled from "@/components/Main/Main.styled";
+import TransactionEntryForm from "@/components/TransactionEntryForm";
+import { useState } from "react";
+import Modal from "@/components/Modal";
 
 export default function TransactionDetailsPage({
   transactions,
   deleteTransaction,
+  handleEditTransaction
 }) {
+  const [showModal, setShowModal] = useState(false);
   const router = useRouter();
   const { id } = router.query;
 
@@ -24,13 +28,20 @@ export default function TransactionDetailsPage({
 
   return (
     <>
-      <Header title="Details" />
-      <Styled.Main>
+     
+      <Modal showModal={showModal}>
+      <TransactionEntryForm  setShowModal={setShowModal}
+        currentTransaction={currentTransaction} 
+        updateTransactions={handleEditTransaction}
+        id={id}
+        mode="edit"/>
+        </Modal>
         <TransactionDetails
+          showModal={showModal} setShowModal={setShowModal}
           currentTransaction={currentTransaction}
           deleteTransaction={() => deleteTransaction(id)}
         />
-      </Styled.Main>
+      
     </>
   );
 }
